Introducing TestRenderer trait for implementing custom renderers
The DefaultTestReporter now has a single responsibility
