+++ test/shared/src/main/scala/zio/test/TestRenderer.scala
@@ -1,0 +1,14 @@
+package zio.test
+
+import zio.UIO
+import zio.duration.Duration
+import zio.test.reflect.Reflect.EnableReflectiveInstantiation
+
+@EnableReflectiveInstantiation
+trait TestRenderer[-E] {
+  def render(
+    executedSpec: ExecutedSpec[E],
+    annotationRenderer: TestAnnotationRenderer
+  ): UIO[Seq[RenderedResult[String]]]
+  def renderStats(duration: Duration, executedSpec: ExecutedSpec[E]): UIO[String]
+}
